requirements.txt
SQLAlchemy
requests
beautifulsoup4
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
python-multipart
